Isolando logica para conectar no Banco de Dados
